페이징

+++ src/main/java/com/manager/freelancer/manager/model/vo/Member.java
@@ -33,7 +33,7 @@ public class Member {
 	private int freeContactTime1;
 	private int freeContactTime2;
 	
-	private int freelancerAccountNo;
+	private long freelancerAccountNo;
 	private String freelancerBankName;
 	
 }
